Rename Backbone Forms overrides
